Use `int`s as scores for `ScoreSupplier`
__Motivation__

`ScoreSupplier#score()` returns a `float` today with an intent to have precisions for slightly varying scores. However floating point arithmentic adds more complexity in score computations. Instead we can use `int`s and use bigger scores for precision (eg: multiply by `1000`)

__Modification__

Make `ScoreSupplier#score()` return an `int`.

__Result__

Simpler score calculations.

+++ servicetalk-http-netty/src/main/java/io/servicetalk/http/netty/LoadBalancedStreamingHttpConnection.java
@@ -158,7 +158,7 @@ final class LoadBalancedStreamingHttpConnection implements FilterableStreamingHt
     }
 
     @Override
-    public float score() {
+    public int score() {
         return filteredConnection.score();
     }
 }
